moved  create_folders and get_args method to utils
